# openapi.yaml
openapi: 3.0.3
info:
  title: Ship Monitor Cloud Auth API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the Gin route registrations and handlers in this service.
    Most endpoints use a cookie-based session named `ship_session`.
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - shipSession: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "201":
          description: User registered
        "400":
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and create a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful. A `ship_session` cookie is set.
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
                example: ship_session=abc123; Path=/; Max-Age=86400; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh the current authenticated session
      responses:
        "200":
          description: Session refreshed. A fresh session cookie is set.
          headers:
            Set-Cookie:
              description: Refreshed session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Log out and clear the session cookie
      security:
        - shipSession: []
        - {}
      responses:
        "200":
          description: Logged out. Session cookie is cleared.
          headers:
            Set-Cookie:
              description: Expired session cookie
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/me:
    get:
      tags: [Users]
      summary: Get the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID
      description: The authenticated user may only fetch their own user record.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "404":
          description: User not found or not accessible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/set-password:
    post:
      tags: [Users]
      summary: Change the current user's password
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordRequest"
      responses:
        "200":
          description: Password changed
        "400":
          description: Invalid request or password rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found or not accessible
  /api/users/{id}/set-email:
    post:
      tags: [Users]
      summary: Change the current user's email
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetEmailRequest"
      responses:
        "200":
          description: Email changed
        "400":
          description: Invalid request or email rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/users/start-email-confirmation:
    post:
      tags: [Users]
      summary: Start email confirmation for the current user
      responses:
        "200":
          description: Confirmation started
        "304":
          description: Email is already confirmed
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/confirm-email/{token}:
    post:
      tags: [Users]
      summary: Confirm the current user's email using a token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email confirmed
        "304":
          description: Email is already confirmed
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v2/devices/{id}/state/{state}:
    get:
      tags: [Devices]
      summary: Get the current or historical state records for a device
      parameters:
        - $ref: "#/components/parameters/DeviceIdAsId"
        - name: state
          in: path
          required: true
          schema:
            type: string
        - name: history
          in: query
          required: false
          description: Number of historical records to return. If omitted, only the current/default record set is requested.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Device state records
          content:
            application/json:
              schema:
                type: object
                required: [result]
                properties:
                  result:
                    type: array
                    items:
                      $ref: "#/components/schemas/StateRecord"
        "400":
          description: Invalid history query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v2/devices/{id}/command:
    post:
      tags: [Devices]
      summary: Send a command to a device
      parameters:
        - $ref: "#/components/parameters/DeviceIdAsId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCommandRequest"
      responses:
        "200":
          description: Command accepted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    shipSession:
      type: apiKey
      in: cookie
      name: ship_session
      description: Session cookie set by `POST /api/auth/login`.
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      description: User UUID
      schema:
        $ref: "#/components/schemas/UUID"
...
    DeviceIdAsId:
      name: id
      in: path
      required: true
      description: Device UUID
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    Unauthorized:
      description: Authentication required or session is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
      example: 80fcdeff-0bc3-4f93-a2b0-00632344ac88
    Timestamp:
      type: string
      format: date-time
      example: "2026-07-09T19:00:37.607709Z"
...
    ErrorResponse:
      type: object
      required: [details]
      properties:
        details:
          type: string
        messageForArmen:
          type: string
          nullable: true
...
    RegisterUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 5
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    User:
      type: object
      required: [id, name, email, emailVerified, blocked, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        blocked:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    UserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    Session:
      type: object
      required: [userId, email]
      properties:
        userId:
          $ref: "#/components/schemas/UUID"
        email:
          type: string
          format: email
    SetPasswordRequest:
      type: object
      required: [oldPassword, password]
      properties:
        oldPassword:
          type: string
          format: password
        password:
          type: string
          format: password
    SetEmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
    SendCommandRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
        args:
          type: object
          additionalProperties: true
    StateRecord:
      type: object
      required: [deviceId, state, timestamp, value]
      properties:
        deviceId:
          type: string
        state:
          type: string
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        value:
          description: State value. Type depends on the requested state.